[backend] refactor(organization): merge helper methods into OrganizationDomain (#2339)
